pyproject.toml
[project]
name = "hierarchical-safety-governor"
version = "0.1.0"
description = "A hierarchical safety governor for AI agent systems"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyautogen[openai]==0.9.1",
    "gymnasium>=0.29.0",
    "pyyaml>=6.0",
    "duckdb>=0.10.0",
    "inspect-ai>=0.3.0",
    "numpy>=1.24.0",
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[project.scripts]
safety-governor = "scripts.run:main"
safety-governor-ui = "scripts.stream_ui:main"
safety-governor-analyze = "scripts.analyze:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/safety_governor"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/scripts",
    "/configs",
    "/docs",
    "/tests",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]